+++ infra/modules/worklytics-proxy-connection/main.tf
@@ -22,9 +22,11 @@ module "generic" {
 }
 
 output "todo" {
-  value = module.generic.todo
+  value       = module.generic.todo
+  description = "[DEPRECATED - use todo_content output instead. TODO: remove in 0.7]"
 }
 
 output "next_todo_step" {
-  value = module.generic.next_todo_step
+  value       = module.generic.next_todo_step
+  description = "[DEPRECATED - todo ordering now handled at root module level via todo_content stage indices. TODO: remove in 0.7]"
 }